requirements.txt
asgiref==3.9.1
black==25.1.0
cfgv==3.4.0
click==8.2.1
coverage==7.10.4
distlib==0.4.0
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
factory_boy==3.3.3
Faker==37.5.3
filelock==3.19.1
identify==2.6.13
iniconfig==2.1.0
isort==6.0.1
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.3.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
PyYAML==6.0.2
ruff==0.12.9
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.34.0
